pyproject.toml
[tool.poetry]
name = "riscv-simenv"
version = "0.8.0"
description = "A tool to reliably pack and spawn a RISC-V workload (app) for simulation."
authors = ["Jiayang Li <jli95@ncsu.edu>"]
license = "Other/Proprietary License"
repository = "https://github.com/s117/riscv-simenv.git"
readme = "README.md"
keywords = ["riscv", "simenv", "anycore"]
classifiers = [
    "Operating System :: POSIX :: Linux",
    "Development Status :: 3 - Alpha",
]
packages = [
    { include = "SimEnvControl" },
    { include = "SyscallAnalysis" },
]
exclude = [
    "debug_entry",
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0.1"
click = "^8.1.7"
pyparsing = "^3.1.1"
bashlex = "^0.18"
fuzzywuzzy = "^0.18.0"
natsort = "^8.4.0"
levenshtein = "^0.24.0"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
coverage = "^7.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
riscv-simenv = "SimEnvControl.sim_env_cli:cli"
